# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API Mongo Project",
        "title": "API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/alumni": {
            "get": {
                "description": "Mengambil daftar semua Alumni dari database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alumni"
                ],
                "summary": "Dapatkan semua Alumni",
                "parameters": [
                    {
                        "description": "Data Alumni",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Alumni"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Alumni"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Mengambil daftar semua Alumni dari database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alumni"
                ],
                "summary": "Dapatkan semua Alumni",
                "parameters": [
                    {
                        "description": "Data Alumni",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Alumni"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Alumni"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/alumni/:nim": {
            "get": {
                "description": "Mengambil daftar semua Alumni dari database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alumni"
                ],
                "summary": "Dapatkan semua Alumni",
                "parameters": [
                    {
                        "description": "Data Alumni",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Alumni"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Alumni"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Mengambil daftar semua Alumni dari database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alumni"
                ],
                "summary": "Dapatkan semua Alumni",
                "parameters": [
                    {
                        "description": "Data Alumni",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Alumni"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Alumni"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Mengambil daftar semua Alumni dari database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alumni"
                ],
                "summary": "Dapatkan semua Alumni",
                "parameters": [
                    {
                        "description": "Data Alumni",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Alumni"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Alumni"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Alumni": {
            "type": "object",
            "properties": {
                "angkatan": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id_fakultas": {
                    "type": "integer"
                },
                "id_prodi": {
                    "type": "integer"
                },
                "id_sumber": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                },
                "nim": {
                    "type": "string"
                },
                "sumber": {
                    "type": "string"
                },
                "tahun_lulus": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "error": {
                    "type": "string",
                    "example": "Failed to fetch users"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}